backend/requirements.txt
Django==4.1.7
Pillow==9.4.0
numpy==1.24.2
opencv-python==4.7.0.72
torch==2.0.0
torchvision==0.15.1
ultralytics==8.0.20
python-dotenv==1.0.0
djangorestframework==3.14.0
django-cors-headers==3.14.0
seaborn==0.12.2
matplotlib==3.7.1
pandas==2.0.0
